Combined site_templates into site_projects
