--- services/nginx/templates/templates.xithrius.cloud.conf.template
@@ -1,8 +1,0 @@
-server {
-    listen 80;
-    server_name templates.xithrius.cloud;
-
-    location / {
-        proxy_pass http://site_templates:8003;
-    }
-}
