--- site_templates/.eslintrc.json
@@ -1,15 +1,0 @@
-{
-  "extends": [
-    "next/core-web-vitals",
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
-  "parserOptions": {
-    "project": "**/tsconfig.json"
-  },
-  "rules": {
-    "@next/next/no-img-element": "off"
-  }
-}
