--- site_templates/.prettierignore
@@ -1,2 +1,0 @@
-node_modules/**
-.next/**
