--- site_templates/next.config.js
@@ -1,8 +1,0 @@
-/** @type {import('next').NextConfig} */
-module.exports = {
-  reactStrictMode: true,
-  swcMinify: true,
-  images: {
-    domains: ["opengraph.githubassets.com"],
-  },
-};
