--- site_templates/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "site_templates",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "next": "^12.2.3",
-    "next-themes": "^0.2.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "autoprefixer": "^10.4.7",
-    "eslint": "^8.19.0",
-    "eslint-config-next": "^12.2.2",
-    "postcss": "^8.4.14",
-    "prettier": "^2.7.1",
-    "prettier-plugin-tailwindcss": "^0.1.13",
-    "tailwindcss": "^3.1.6",
-    "@types/node": "17.0.35",
-    "@typescript-eslint/eslint-plugin": "^5.39.0",
-    "@typescript-eslint/eslint-plugin-tslint": "^5.39.0",
-    "@typescript-eslint/parser": "^5.39.0",
-    "cssnano": "^5.1.12",
-    "typescript": "^4.8.4"
-  }
-}
