--- site_templates/pages/_app.jsx
@@ -1,10 +1,0 @@
-import "../styles/globals.css";
-import { ThemeProvider } from "next-themes";
-
-export default function Site({ Component, pageProps }) {
-  return (
-    <ThemeProvider defaultTheme="system">
-      <Component {...pageProps} />
-    </ThemeProvider>
-  );
-}
